+++ src/components/form/Label.tsx
@@ -1,35 +1,36 @@
 import React from "react";
 
-import classnames from "classnames";
+import { StyledLabel, StyledText } from "./Label.styled";
 
-import { makeVariants } from "../../utils/index";
+type LabelProps = React.ComponentProps<typeof StyledLabel>;
 
-import styles from "./Label.module.scss";
-
-interface Props {
+interface Props extends Partial<Pick<LabelProps, "css" | "className">> {
     label?: React.ReactNode;
     variant?: "big";
-    className?: string;
     children: React.ReactNode;
 }
 
 const Label: React.FC<Props> = (props) => {
-    const v = makeVariants(props.variant);
+    const { label, variant, children, ...rest } = props;
 
-    const labelCls = classnames(styles.text, {
-        [styles.textBig as string]: v.includes("big"),
-    });
-
-    const label = props.label ? <div className={labelCls}>{props.label}</div> : null;
-
-    const rootCls = classnames(props.className, styles.label);
+    const labelElement = label
+        ? (
+            <StyledText {...(variant ? { variant } : undefined)}>
+                {label}
+            </StyledText>
+        )
+        : null;
 
     return (
-        <label className={rootCls}>
-            {label}
-            {props.children}
-        </label>
+        <StyledLabel {...rest}>
+            {labelElement}
+            {children}
+        </StyledLabel>
     );
 };
 
 export { Label };
+
+export type {
+    Props as LabelProps,
+};