+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [EZEZ Changelog](https://ezez.dev/guidelines/changelog/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+### Changed
+- `Label` converted to stitches
+
 ## [0.27.11] - 2025-05-19
 ### Breaking
 - Removed `BackgroundClassName` and `ValueClassName` from `Progress` component, that shouldn't be exported in the first place