+++ CHANGELOG.md
@@ -3,11 +3,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [EZEZ Changelog](https://ezez.dev/guidelines/changelog/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [UNRELEASED]
+## [0.28.0] - 2025-08-16
+### Added
+- `ColorPicker` component
 ### Changed
 - `Label` converted to stitches
 - `Toggle` converted to stitches
 - `Choice` converted to stitches
+### Dev
+- updated some deps
 
 ## [0.27.11] - 2025-05-19
 ### Breaking